+++ Sources/StreamChat/Workers/CurrentUserUpdater.swift
@@ -206,13 +206,14 @@ class CurrentUserUpdater: Worker, @unchecked Sendable {
     }
 
     func setPushPreference(
-        _ preference: PushPreferenceRequestPayload,
+        _ preference: PushPreferenceInput,
         completion: @escaping @Sendable (Result<PushPreference, Error>) -> Void
     ) {
-        apiClient.request(endpoint: .pushPreferences([preference])) { [weak self] result in
+        let request = UpsertPushPreferencesRequest(preferences: [preference])
+        apiClient.request(endpoint: .updatePushNotificationPreferences(upsertPushPreferencesRequest: request)) { [weak self] result in
             switch result {
             case let .success(response):
-                guard let currentUserPushPref = response.userPreferences.asModel().first else {
+                guard let userPreference = response.userPreferences.values.first else {
                     completion(.failure(ClientError.CurrentUserDoesNotExist()))
                     return
                 }
@@ -223,14 +224,11 @@ class CurrentUserUpdater: Worker, @unchecked Sendable {
                     }
                     let savedDTO = try session.savePushPreference(
                         id: currentUserDTO.user.id,
-                        payload: .init(
-                            chatLevel: currentUserPushPref.level.rawValue,
-                            disabledUntil: currentUserPushPref.disabledUntil
-                        )
+                        payload: userPreference
                     )
                     currentUserDTO.pushPreference = savedDTO
                 }
-                completion(.success(currentUserPushPref))
+                completion(.success(userPreference))
             case let .failure(error):
                 completion(.failure(error))
             }